Funnel Success
This project required building a conversion funnel, calculating retention rates, as well as organizing ad showing documentation of steps throughout the spreadsheet. The finished project was required to be formatted as though presenting to a business, making it as understandable and easily navigate-able as possible. Data was prepared by narrowing it down to only purchased activity, calculated first purchase dates, created an event month and cohort age column.