NYC AirBnB
This was my first project in the TripleTen Business Intelligence Analytics Bootcamp. It required cleaning the raw data, building pivot tables, as well as documenting and formatting my work in a consistent and easy-to-understand way. With the provided data I was able to determine that studios and 1 bedrooms are the most popular bedroom size for vacation rentals. Also learned the neighborhoods that were top performing for vacation rentals.