Apps and Reviews
This project was built using Power BI and due to the size of the document, it has been submitted as a folder of screenshots for easiest review. The first page provides basic statistics on the types of apps that are listed in the data. Next is a summary of the review data, both an average value of the helpful reviews and the average rating of the developer answered reviews.